Split Client interfaces and move actual clients into dedicated a package
