+++ sdk-core/src/main/java/care/data4life/sdk/client/Fhir4RecordClient.kt
@@ -14,23 +14,28 @@
  * contact D4L by email to help@data4life.care.
  */
 
-package care.data4life.sdk.resource
+package care.data4life.sdk.client
 
+import care.data4life.sdk.SdkContract
 import care.data4life.sdk.auth.AuthContract
 import care.data4life.sdk.call.CallHandler
 import care.data4life.sdk.call.Callback
-import care.data4life.sdk.call.DataRecord
+import care.data4life.sdk.call.Fhir4Record
 import care.data4life.sdk.call.Task
+import care.data4life.sdk.model.DownloadType
 import care.data4life.sdk.record.RecordContract
+import care.data4life.sdk.resource.Fhir4Attachment
+import care.data4life.sdk.resource.Fhir4Resource
 import care.data4life.sdk.tag.Annotations
 import io.reactivex.Single
 import org.threeten.bp.LocalDate
 
-internal class DataRecordClient(
+internal class Fhir4RecordClient(
     private val userService: AuthContract.UserService,
     private val recordService: RecordContract.Service,
     private val handler: CallHandler
-) : ResourceContract.Client {
+) : SdkContract.Fhir4RecordClient {
+
     private fun <T> executeOperationFlow(
         operation: (userId: String) -> Single<T>,
         callback: Callback<T>
@@ -41,44 +46,46 @@ internal class DataRecordClient(
         return handler.executeSingle(flow, callback)
     }
 
-    override fun create(
-        resource: ResourceContract.DataResource,
+    override fun <T : Fhir4Resource> create(
+        resource: T,
         annotations: Annotations,
-        callback: Callback<DataRecord<ResourceContract.DataResource>>
+        callback: Callback<Fhir4Record<T>>
     ): Task = executeOperationFlow(
         { userId -> recordService.createRecord(userId, resource, annotations) },
         callback
     )
 
-    override fun update(
+    override fun <T : Fhir4Resource> update(
         recordId: String,
-        resource: ResourceContract.DataResource,
+        resource: T,
         annotations: Annotations,
-        callback: Callback<DataRecord<ResourceContract.DataResource>>
+        callback: Callback<Fhir4Record<T>>
     ): Task = executeOperationFlow(
         { userId -> recordService.updateRecord(userId, recordId, resource, annotations) },
         callback
     )
 
-    override fun fetch(
+    override fun <T : Fhir4Resource> fetch(
         recordId: String,
-        callback: Callback<DataRecord<ResourceContract.DataResource>>
+        callback: Callback<Fhir4Record<T>>
     ): Task = executeOperationFlow(
-        { userId -> recordService.fetchDataRecord(userId, recordId) },
+        { userId -> recordService.fetchFhir4Record(userId, recordId) },
         callback
     )
 
-    override fun search(
+    override fun <T : Fhir4Resource> search(
+        resourceType: Class<T>,
         annotations: Annotations,
         startDate: LocalDate?,
         endDate: LocalDate?,
         pageSize: Int,
         offset: Int,
-        callback: Callback<List<DataRecord<ResourceContract.DataResource>>>
+        callback: Callback<List<Fhir4Record<T>>>
     ): Task = executeOperationFlow(
         { userId ->
-            recordService.fetchDataRecords(
+            recordService.fetchFhir4Records(
                 userId,
+                resourceType,
                 annotations,
                 startDate,
                 endDate,
@@ -89,11 +96,20 @@ internal class DataRecordClient(
         callback
     )
 
-    override fun count(
+    override fun <T : Fhir4Resource> download(
+        recordId: String,
+        callback: Callback<Fhir4Record<T>>
+    ): Task = executeOperationFlow(
+        { userId -> recordService.downloadFhir4Record(recordId, userId) },
+        callback
+    )
+
+    override fun <T : Fhir4Resource> count(
+        resourceType: Class<T>,
         annotations: Annotations,
         callback: Callback<Int>
     ): Task = executeOperationFlow(
-        { userId -> recordService.countDataRecords(DataResource::class.java, userId, annotations) },
+        { userId -> recordService.countFhir4Records(resourceType, userId, annotations) },
         callback
     )
 
@@ -102,6 +118,26 @@ internal class DataRecordClient(
         callback: Callback<Boolean>
     ): Task = executeOperationFlow(
         { userId -> recordService.deleteRecord(userId, recordId).toSingle { true } },
+        callback
+    )
+
+    override fun downloadAttachment(
+        recordId: String,
+        attachmentId: String,
+        type: DownloadType,
+        callback: Callback<Fhir4Attachment>
+    ): Task = executeOperationFlow(
+        { userId -> recordService.downloadFhir4Attachment(recordId, attachmentId, userId, type) },
+        callback
+    )
+
+    override fun downloadAttachments(
+        recordId: String,
+        attachmentIds: List<String>,
+        type: DownloadType,
+        callback: Callback<List<Fhir4Attachment>>
+    ): Task = executeOperationFlow(
+        { userId -> recordService.downloadFhir4Attachments(recordId, attachmentIds, userId, type) },
         callback
     )
 }
